moved validation to main process object
